+++ Contacts.Web.API/Controllers/SkillsController.cs
@@ -9,6 +9,7 @@ using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Contacts.Web.API.Models.ContextConfiguration;
 using System.Security.Claims;
 using System.Net;
+using Contacts.Web.API.Models.DTO;
 
 namespace Contacts.Web.API.Controllers
 {
@@ -26,34 +27,44 @@ namespace Contacts.Web.API.Controllers
 
         // GET: api/Skills
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<Skill>>> GetSkillItems() 
+        public async Task<ActionResult<IEnumerable<SkillDTO>>> GetSkillItems()
         {
-            return await _context.SkillItems.ToListAsync();
+            List<SkillDTO> result = new List<SkillDTO>();
+
+            List<Skill> skillsList = await _context.SkillItems.Include(s => s.Contact).ToListAsync();
+            foreach (var skill in skillsList)
+            {
+                result.Add(new SkillDTO() { Id = skill.Id, idContact = skill.Contact != null ? skill.Contact.Id : 0, Level = skill.Level, Name = skill.Name });
+            }
+
+            return result;
         }
 
         // GET: api/Skills/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<Skill>> GetSkill([FromQuery] long id)
+        public async Task<ActionResult<SkillDTO>> GetSkill([FromRoute] long id)
         {
-            var skill = await _context.SkillItems.FindAsync(id);
+            Skill skill = _context.SkillItems.Include(s => s.Contact).Where(s => s.Id == id).SingleOrDefaultAsync().Result;
 
             if (skill == null)
             {
                 return NotFound();
             }
 
-            return skill;
+            return new SkillDTO() { Id = skill.Id, idContact = skill.Contact.Id, Level = skill.Level, Name = skill.Name };
         }
 
         // PUT: api/Skills/5
         // To protect from overposting attacks, enable the specific properties you want to bind to, for
         // more details, see https://go.microsoft.com/fwlink/?linkid=2123754.
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutSkill([FromQuery] long id, [FromBody] Skill skill)
+        public async Task<IActionResult> PutSkill([FromRoute] long id, [FromBody] SkillDTO skillDTO)
         {
-            Skill skillWithContact = _context.SkillItems.Include(s => s.Contact).Where(s => s.Id == id).SingleOrDefaultAsync().Result;
+            Skill skillWithContact = _context.SkillItems.Include(s => s.Contact).Where(s => s.Id == skillDTO.Id).SingleOrDefaultAsync().Result;
 
             if (skillWithContact == null) return StatusCode((int)HttpStatusCode.NotFound, $"Cannot find a skill with id {id}...");
+
+            if (skillWithContact.Contact.Id != skillDTO.idContact) return StatusCode((int)HttpStatusCode.BadRequest, $"You can't update the contact of the skill...");
 
             // Authorization
             var identity = HttpContext.User.Identity as ClaimsIdentity;
@@ -61,12 +72,15 @@ namespace Contacts.Web.API.Controllers
             string currentEmailConnected = claims[0].Value;
             if (skillWithContact.Contact?.Email.ToLower() != currentEmailConnected.ToLower()) return StatusCode((int)HttpStatusCode.Unauthorized, "Sorry! You can't change data of a skill that isn't yours...");
 
-            if (id != skill.Id)
+            if (id != skillWithContact.Id)
             {
                 return BadRequest();
             }
 
-            _context.Entry(skill).State = EntityState.Modified;
+            // update fields
+            skillWithContact.Level = skillDTO.Level;
+            skillWithContact.Name = skillDTO.Name;
+            _context.Entry(skillWithContact).State = EntityState.Modified;
 
             try
             {
@@ -91,8 +105,15 @@ namespace Contacts.Web.API.Controllers
         // To protect from overposting attacks, enable the specific properties you want to bind to, for
         // more details, see https://go.microsoft.com/fwlink/?linkid=2123754.
         [HttpPost]
-        public async Task<ActionResult<Skill>> PostSkill([FromBody] Skill skill)
+        public async Task<ActionResult<Skill>> PostSkill([FromBody] SkillDTO skillDTO)
         {
+            if (skillDTO.idContact == 0) return StatusCode((int)HttpStatusCode.BadRequest, $"A skill must have a contact...");
+            Contact contact = _context.ContactItems.FindAsync(skillDTO.idContact).Result;
+
+            if (contact == null) return StatusCode((int)HttpStatusCode.NotFound, $"We can't find a contact with id {skillDTO.idContact}...");
+
+            Skill skill = new Skill() { Level = skillDTO.Level, Name = skillDTO.Name };
+            skill.Contact = contact; // contact association
             _context.SkillItems.Add(skill);
             await _context.SaveChangesAsync();
 
@@ -101,13 +122,20 @@ namespace Contacts.Web.API.Controllers
 
         // DELETE: api/Skills/5
         [HttpDelete("{id}")]
-        public async Task<ActionResult<Skill>> DeleteSkill([FromQuery] long id)
+        public async Task<ActionResult<Skill>> DeleteSkill([FromRoute] long id)
         {
             var skill = await _context.SkillItems.FindAsync(id);
             if (skill == null)
             {
                 return NotFound();
             }
+
+            // Authorization
+            Skill skillWithContact = _context.SkillItems.Include(s => s.Contact).Where(s => s.Id == id).SingleOrDefaultAsync().Result;
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
+            IList<Claim> claims = identity.Claims.ToList();
+            string currentEmailConnected = claims[0].Value;
+            if (skillWithContact.Contact?.Email.ToLower() != currentEmailConnected.ToLower()) return StatusCode((int)HttpStatusCode.Unauthorized, "Sorry! You can't delete a skill that isn't yours...");
 
             _context.SkillItems.Remove(skill);
             await _context.SaveChangesAsync();
